task1_edge_recycling_classifier/requirements.txt
# Edge AI Recyclable Classifier - Requirements

# Core Deep Learning
tensorflow>=2.13.0,<2.16.0
tensorflow-lite>=2.13.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning Utilities
scikit-learn>=1.3.0

# Image Augmentation
imgaug>=0.4.0

# Progress Bars
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Model Optimization (optional)
# tensorflow-model-optimization>=0.7.0

# For Raspberry Pi deployment (lightweight alternative)
# Install this instead on Raspberry Pi:
# tflite-runtime>=2.13.0

# Additional utilities
opencv-python>=4.8.0  # For image processing and camera
requests>=2.31.0       # For dataset downloading
